testcases fixed
