+++ server/src/main/java/com/flipkart/varadhi/auth/DefaultAuthorizationProvider.java
@@ -12,7 +12,7 @@ import com.flipkart.varadhi.services.IamPolicyService;
 import com.flipkart.varadhi.spi.ConfigFileResolver;
 import com.flipkart.varadhi.spi.authz.AuthorizationOptions;
 import com.flipkart.varadhi.spi.authz.AuthorizationProvider;
-import com.flipkart.varadhi.spi.db.IamPolicyMetaStore;
+import com.flipkart.varadhi.spi.db.IamPolicy.IamPolicyOperations;
 import com.flipkart.varadhi.spi.db.MetaStore;
 import com.flipkart.varadhi.spi.db.MetaStoreOptions;
 import com.flipkart.varadhi.spi.db.MetaStoreProvider;
@@ -62,9 +62,9 @@ public class DefaultAuthorizationProvider implements AuthorizationProvider, Auto
             metaStoreProvider.init(options);
             MetaStore store = metaStoreProvider.getMetaStore();
 
-            if (!(store instanceof IamPolicyMetaStore)) {
+            if (!(store instanceof IamPolicyOperations)) {
                 throw new IllegalStateException(
-                    String.format("Provider %s must implement IamPolicyMetaStore", options.getProviderClassName())
+                    String.format("Provider %s must implement IamPolicyOperations", options.getProviderClassName())
                 );
             }
 
@@ -72,7 +72,7 @@ public class DefaultAuthorizationProvider implements AuthorizationProvider, Auto
                 "Successfully initialized authorization service with provider: {}",
                 options.getProviderClassName()
             );
-            return new IamPolicyService(store, (IamPolicyMetaStore)store);
+            return new IamPolicyService(store, (IamPolicyOperations)store);
         } catch (Exception e) {
             cleanupProvider();
             throw new IllegalStateException("Failed to initialize authorization service", e);
