+++ server/src/main/java/com/flipkart/varadhi/verticles/webserver/WebServerVerticle.java
@@ -16,7 +16,7 @@ import com.flipkart.varadhi.produce.otel.ProducerMetricHandler;
 import com.flipkart.varadhi.produce.services.ProducerService;
 import com.flipkart.varadhi.services.*;
 import com.flipkart.varadhi.spi.ConfigFileResolver;
-import com.flipkart.varadhi.spi.db.IamPolicyMetaStore;
+import com.flipkart.varadhi.spi.db.IamPolicy.IamPolicyOperations;
 import com.flipkart.varadhi.spi.db.MetaStore;
 import com.flipkart.varadhi.spi.services.MessagingStackProvider;
 import com.flipkart.varadhi.spi.services.Producer;
@@ -170,7 +170,7 @@ public class WebServerVerticle extends AbstractVerticle {
             null :
             configuration.getAuthorization().getProviderClassName();
         boolean isDefaultProvider = DefaultAuthorizationProvider.class.getName().equals(authProviderName);
-        boolean isIamPolicyStore = metaStore instanceof IamPolicyMetaStore;
+        boolean isIamPolicyStore = metaStore instanceof IamPolicyOperations;
         //TODO::Validate below specifically w.r.to lean deployment.
         // enable IamPolicy Routes, if
         // 1. provider class name is DefaultAuthorizationProvider, and
@@ -181,7 +181,7 @@ public class WebServerVerticle extends AbstractVerticle {
                 routes.addAll(
                     new IamPolicyHandlers(
                         projectService,
-                        new IamPolicyService(metaStore, (IamPolicyMetaStore)metaStore)
+                        new IamPolicyService(metaStore, (IamPolicyOperations)metaStore)
                     ).get()
                 );
             } else {
