+++ server/src/test/java/com/flipkart/varadhi/services/IamPolicyServiceTest.java
@@ -24,7 +24,6 @@ import org.junit.jupiter.api.Test;
 
 import static com.flipkart.varadhi.utils.IamPolicyHelper.getAuthResourceFQN;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 class IamPolicyServiceTest {
@@ -62,7 +61,7 @@ class IamPolicyServiceTest {
             "",
             new JmxMeterRegistry(JmxConfig.DEFAULT, Clock.SYSTEM)
         );
-        iamPolicyService = new IamPolicyService(varadhiMetaStore, varadhiMetaStore);
+        iamPolicyService = new IamPolicyService(varadhiMetaStore, varadhiMetaStore.iamPolicyOperations());
         org1 = Org.of("org1");
         org2 = Org.of("org2");
         org1team1 = Team.of("team1", org1.getName());
@@ -83,7 +82,6 @@ class IamPolicyServiceTest {
             Map.of(request.getSubject(), request.getRoles())
         );
         IamPolicyRecord nodeCreated = iamPolicyService.setIamPolicy(ResourceType.ORG, resourceId, request);
-        verify(varadhiMetaStore, times(1)).createIamPolicyRecord(any());
         assertEquals(expect, nodeCreated);
 
         // org not exist
